+++ docs/examples/solana/02.swaps.example.ts
@@ -1,4 +1,4 @@
-import { solanaInstructionDecoder, solanaPortalSource } from '@subsquid/pipes/solana'
+import { solanaInstructionDecoder, solanaPortalStream } from '@subsquid/pipes/solana'
 
 import * as meteoraDamm from './abi/meteora-damm/index.js'
 import * as meteoraDlmm from './abi/meteora-dlmm/index.js'
@@ -10,7 +10,7 @@ import * as raydiumCpmm from './abi/raydium-cpmm/index.js'
 const from = '340,000,000'
 
 async function cli() {
-  const stream = solanaPortalSource({
+  const stream = solanaPortalStream({
     id: 'solana-swaps',
     portal: 'https://portal.sqd.dev/datasets/solana-mainnet',
     outputs: {